atiadlhandle
